Simplified vendor link card
