+++ web/src/components/vendors/VendorLinkCard.vue
@@ -3,74 +3,68 @@
     v-if="isLoading || isNil(vendor)"
     type="image"
   />
-  <v-card
+  <SimpleCard
     v-else
     class="mb-5"
+    :title="vendor.name"
     @click="goToVendor"
   >
-    <v-card-title>{{ vendor.name }} <small style="color: #888"></small></v-card-title>
-    <!--  <v-card-subtitle>{{ vendor.vendorId }}</v-card-subtitle> -->
-    <v-card-text>
-      <div class="d-flex mb-3">
-        <v-icon
-          class="mt-2"
-          size="40"
-          color="#7A9A01"
-        >
-          mdi-store
-        </v-icon>
-        <div class="ml-2 text-subtitle-1">
-          <strong>Vendor ID: </strong><br />{{ vendor.vendorId }}
-        </div>
-      </div>
-
-      <div class="d-flex mb-3">
-        <v-icon
-          class="mt-2"
-          size="40"
-          color="#7A9A01"
-        >
-          mdi-map
-        </v-icon>
-        <div class="ml-2 text-subtitle-1">
-          <strong>Address:</strong>
-          <div
-            v-for="(line, index) in formatAddressLines(vendor)"
-            :key="index"
-          >
-            {{ line }}
-          </div>
-        </div>
-      </div>
-
-      <div class="d-flex">
-        <v-icon
-          class="mt-2"
-          size="40"
-          color="#7A9A01"
-        >
-          mdi-handshake
-        </v-icon>
-        <div class="ml-2 text-subtitle-1">
-          <strong>Programs: </strong><br />
-
-          <div
-            v-for="(program, index) in programs"
-            :key="index"
-          >
-            {{ program.name }}
-          </div>
-        </div>
-      </div>
-    </v-card-text>
-    <v-card-text>
-      <v-btn
-        block
-        text="Go To Vendor"
-        @click="goToVendor"
+    <div class="d-flex mb-3">
+      <v-icon
+        class="mt-2"
+        size="40"
+        color="#7A9A01"
+        icon="mdi-store"
       />
-    </v-card-text>
-  </v-card>
+      <div class="ml-2 text-subtitle-1">
+        <strong>Vendor ID: </strong><br />{{ vendor.vendorId }}
+      </div>
+    </div>
+
+    <div class="d-flex mb-3">
+      <v-icon
+        class="mt-2"
+        size="40"
+        color="#7A9A01"
+        icon="mdi-map"
+      />
+      <div class="ml-2 text-subtitle-1">
+        <strong>Address:</strong>
+        <div
+          v-for="(line, index) in formatVendorAddressLines(vendor)"
+          :key="index"
+        >
+          {{ line }}
+        </div>
+      </div>
+    </div>
+
+    <div class="d-flex">
+      <v-icon
+        class="mt-2"
+        size="40"
+        color="#7A9A01"
+        icon="mdi-handshake"
+      />
+      <div class="ml-2 text-subtitle-1">
+        <strong>Programs: </strong><br />
+
+        <div
+          v-for="(program, index) in programs"
+          :key="index"
+        >
+          {{ program.name }}
+        </div>
+      </div>
+    </div>
+
+    <v-btn
+      block
+      class="mt-5"
+      text="Go To Vendor"
+      @click="goToVendor"
+    />
+  </SimpleCard>
 </template>
 
 <script setup lang="ts">
@@ -78,8 +72,12 @@ import { computed, toRefs } from "vue"
 import { useRouter } from "vue-router"
 import { isNil } from "lodash"
 
-import useVendor, { Vendor } from "@/use/use-vendor"
+import { formatVendorAddressLines } from "@/utils/format-vendor-address-lines"
+
+import useVendor from "@/use/use-vendor"
 import usePrograms, { ProgramQueryOptions } from "@/use/use-programs"
+
+import SimpleCard from "@/components/common/SimpleCard.vue"
 
 const props = defineProps<{ vendorId: number }>()
 const { vendorId } = toRefs(props)
@@ -99,20 +97,6 @@ const programsQuery = computed<ProgramQueryOptions>(() => {
 })
 
 const { programs } = usePrograms(programsQuery, { skipWatchIf: () => isNil(vendor.value) })
-
-function formatAddressLines(vendor: Vendor): string[] {
-  const parts = []
-
-  if (vendor.addressLine1) {
-    parts.push(vendor.addressLine1)
-  } else if (vendor.addressLine2) {
-    parts.push(vendor.addressLine2)
-  }
-
-  parts.push(`${vendor.addressCity} ${vendor.addressProvince}, ${vendor.addressPostal}`)
-
-  return parts.filter((part) => part.trim() !== "")
-}
 
 function goToVendor() {
   router.push({ name: "vendor/HomePage", params: { vendorId: vendor.value?.vendorId } })
